requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-jose
# passlib[bcrypt]
pwdlib[argon2]
pydantic
python-dotenv
pydantic[email]
python-decouple
python-multipart